spec: don't destroy defaultshard
Reviewed-on: https://gerrit.instructure.com/37685
Tested-by: Jenkins <jenkins@instructure.com>
Product-Review: James Williams  <jamesw@instructure.com>
QA-Review: James Williams  <jamesw@instructure.com>

+++ spec/sharding_spec_helper.rb
@@ -33,7 +33,7 @@ unless SHARDING_ENABLED
   unless CANVAS_RAILS2
     RSpec.configure do |config|
       config.before :all do
-        Shard.default.destroy
+        Shard.default.destroy if Shard.default.is_a?(Shard)
         Shard.default(true)
       end
     end
